[wit-component] add shared-everything linking support
Per https://github.com/WebAssembly/component-model/blob/main/design/mvp/examples/SharedEverythingDynamicLinking.md and https://hackmd.io/IlY4lICRRNy9wQbNLdb2Wg.

This adds a new `component link` subcommand, which resembles `component new` but
accepts an arbitrary number of "dynamic library" modules (as defined by
https://github.com/WebAssembly/tool-conventions/blob/main/DynamicLinking.md),
linking them together to create a component whose type is the union of the
component types found in the input modules (via the usual custom section
representation).  This process can also be invoked programmatically via a new
`Linker` API.

`Linker` analyzes and topologically sorts the input modules, then sythesizes two
additional modules:

- `main`: hosts the component's single memory and function table and exports any
  functions needed to break dependency cycles discovered in the input modules.
  Those functions use `call.indirect` to invoke the real functions, references
  to which are placed in the table by the `init` module.

- `init`: populates the function table as described above, initializes global
  variables per the dynamic linking tool convention, and calls any static
  constructors and/or link-time fixup functions

Finally, `Linker` generates a directed graph of modules and their instantiations
and passes it to `ComponentEncoder` to actually produce the component.

The implementation relies on a few modifications to the existing code in
`ComponentEncoder` and friends.  Specifically, I've expanded the meaning of
"adapter" to include dynamic libraries as well as traditional adapters.  The
former are distiguished from the latter by the presence of a new `LibraryInfo`
struct.  This approach minimized the number of code changes in the interest of
keeping this PR comprehensible, but we will probably want to refactor
`ComponentEncoder` in the future to make the distinction between traditional
adapters and dynamic libraries clearer, possibly generalizing how we analyze and
instantiate modules of various flavors.

I've created a demo which ties everything together, for reference:
https://github.com/dicej/component-linking-demo.  It includes `wasi-sdk` and
`wasi-libc` patches which have not yet been upstreamed; I'm planning to work on
that next.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

skip wasmtime part of linking test under Wasm

Signed-off-by: Joel Dice <joel.dice@fermyon.com>